+++ previous-versions/accounting/RELEASES.md
@@ -401,3 +401,13 @@ Based on:
 - [go v0.29.0] previous-versions/accounting
 ### Releases
 - [Go v0.29.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.29.0 - previous-versions/accounting
+
+## 2023-11-20 01:30:49
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.29.1] previous-versions/accounting
+### Releases
+- [Go v0.29.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.29.1 - previous-versions/accounting